+++ src/content/input-column/content-input-column.component.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import ContentInputRow from '../input-row/content-input-row.component';
+
+const ContentInputColumn = props => (<ContentInputRow asColumn {...props} />);
+export default ContentInputColumn;