+++ src/content/input-row/content-input-row.component.jsx
@@ -6,18 +6,27 @@ import { theme } from '../../constants';
 
 const getErrorStyles = error => (error ? theme.text.error.color : '');
 
+const getContainerDirection = isCol =>
+  (isCol ? theme.container.inputColumn : theme.container.inputRow);
 const Container = styled.section`
   display: flex;
+  flex-direction: ${props => getContainerDirection(props.asColumn)};
   margin-bottom: ${theme.halfGutterWidth};
   @media (max-width: 640px) {
     flex-direction: column;
   }
 `;
 
+const getLabelContainerWidth = isCol =>
+  (isCol ? theme.inputColumn.labelWidth : theme.inputRow.labelWidth);
+const getLabelContainerLineHeight = isCol =>
+  (isCol ? theme.inputColumn.labelContainerLineHeight : theme.inputRow.labelContainerLineHeight);
+
 const LabelContainer = styled.div`
-  width: ${theme.inputRow.labelWidth};
-  min-width: ${theme.inputRow.labelWidth};
+  width: ${props => getLabelContainerWidth(props.asColumn)};
+  min-width: ${props => getLabelContainerWidth(props.asColumn)};
   color: ${props => getErrorStyles(props.error)};
+  line-height: ${props => getLabelContainerLineHeight(props.asColumn)};
 `;
 
 const ValueContainer = styled.div`
@@ -27,8 +36,11 @@ const ValueContainer = styled.div`
   flex: 1 1 auto;
 `;
 
+const getErrorContainerMinHeight = isCol =>
+  (isCol ? theme.inputColumn.errorContainer.height : theme.inputRow.errorContainer.height);
+
 const ErrorContainer = styled.div`
-  min-height: ${theme.inputRow.errorContainer.height};
+  min-height: ${props => getErrorContainerMinHeight(props.asColumn)};
   display: flex;
   align-items: center;
 `;
@@ -38,14 +50,20 @@ const ErrorMessage = styled.span`
   font-size: ${theme.text.error.fontSize};
 `;
 
+const getRequiredIndicatorFontSize = isCol =>
+  (isCol ? theme.inputColumn.requiredIndicator.fontSize : theme.inputRow.requiredIndicator.fontSize);
+
 const RequiredIndicator = styled.span`
   margin-left: ${theme.halfGutterWidth};
-  font-size: ${theme.inputRow.requiredIndicator.fontSize};
+  font-size: ${props => getRequiredIndicatorFontSize(props.asColumn)};
 `;
 
+const getLabelRightMargin = isCol => (isCol ? 0 : theme.gutterWidth);
+const getLabelBottomMargin = isCol => (isCol ? theme.gutterWidth : 0);
+
 const Label = styled.label`
-  margin-right: ${theme.gutterWidth};
-  margin-bottom: 0;
+  margin-right: ${props => getLabelRightMargin(props.asColumn)};
+  margin-bottom: ${props => getLabelBottomMargin(props.asColumn)};
 `;
 
 /**
@@ -67,21 +85,21 @@ const modifyChildren = (child, props) => {
 
 const ContentInputRow = (props) => {
   const {
-    children, error, showError, label, className, required, id,
+    children, error, showError, label, className, required, id, asColumn,
   } = props;
 
   return (
-    <Container className={className} id={id}>
-      <LabelContainer error={error}>
+    <Container asColumn={asColumn} className={className} id={id}>
+      <LabelContainer asColumn={asColumn} error={error}>
         {label &&
-        <Label>{label}
-          {required && <RequiredIndicator>*</RequiredIndicator>}
+        <Label asColumn={asColumn}>{label}
+          {required && <RequiredIndicator asColumn={asColumn}>*</RequiredIndicator>}
         </Label>}
       </LabelContainer>
       <ValueContainer error={error}>
         {React.Children.map(children, child => modifyChildren(child, props))}
         {showError &&
-        <ErrorContainer>
+        <ErrorContainer asColumn={asColumn}>
           {error && <ErrorMessage>{error}</ErrorMessage>}
         </ErrorContainer>}
       </ValueContainer>
@@ -97,6 +115,7 @@ ContentInputRow.propTypes = {
   showError: PropTypes.bool,
   className: PropTypes.string,
   id: PropTypes.string,
+  asColumn: PropTypes.bool,
 };
 
 ContentInputRow.defaultProps = {
@@ -107,6 +126,7 @@ ContentInputRow.defaultProps = {
   showError: true,
   className: '',
   id: '',
+  asColumn: false,
 };
 
 export default ContentInputRow;